+++ src/main/kotlin/com/ekino/oss/jcv/idea/plugin/language/JcvFileType.kt
@@ -3,7 +3,7 @@ package com.ekino.oss.jcv.idea.plugin.language
 import com.intellij.openapi.fileTypes.LanguageFileType
 import javax.swing.Icon
 
-object JcvFileType : LanguageFileType(JcvLanguage) {
+class JcvFileType : LanguageFileType(JcvLanguage) {
 
   override fun getName(): String = "JCV"