+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - **Intention definition** - Add missing `JSON` language to JCV replacement intention configuration.
 - **API usage** - Remove "override-only" api usage and deprecated ones.
+- **File type** - Replaced object by class
 
 ## [3.0.3] - 2025-07-24